backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
cffi==1.15.1
cryptography==38.0.1
pycparser==2.21
django-cors-headers===3.13.0
psycopg2===2.9.5
pytest===7.3.2
pytest-django===4.5.2
Faker===18.10.1
